Bus is now the interface. BasicBus is now the impl.
Also updated comments and code to comply with current styles. Version is also bumped to 2.0.0-SNAPSHOT.
