+++ checkstyle.xml
@@ -51,7 +51,7 @@
         <!-- Checks for Size Violations.                    -->
         <!-- See http://checkstyle.sf.net/config_sizes.html -->
         <module name="LineLength">
-            <property name="max" value="120"/>
+            <property name="max" value="100"/>
         </module>
         <module name="MethodLength"/>
         <module name="ParameterNumber"/>
