+++ library/pom.xml
@@ -20,13 +20,11 @@
   <parent>
     <groupId>com.squareup</groupId>
     <artifactId>otto-parent</artifactId>
-    <version>1.3.4-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
-  <groupId>com.squareup</groupId>
   <artifactId>otto</artifactId>
-  <packaging>jar</packaging>
   <name>Otto</name>
 
   <dependencies>
@@ -47,20 +45,4 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
-
-  <build>
-    <finalName>${binary.prefix}-${project.artifactId}-${project.version}</finalName>
-
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-      </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-checkstyle-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
 </project>
