+++ library/src/main/java/com/squareup/otto/BasicBus.java
@@ -1,0 +1,413 @@
+/*
+ * Copyright (C) 2012 Square, Inc.
+ * Copyright (C) 2007 The Guava Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.squareup.otto;
+
+import java.lang.reflect.InvocationTargetException;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.CopyOnWriteArraySet;
+
+/**
+ * The reference implementation for the Bus interface.  Dispatches events to listeners, and
+ * provides ways for listeners to register themselves.
+ * <p>
+ * BasicBus, by default, enforces that all interactions occur on the main thread.  You can
+ * provide an alternate enforcement by passing a {@link ThreadEnforcer} to the constructor.
+ * <p>
+ * This class is safe for concurrent use.
+ *
+ * @author Cliff Biffle
+ * @author Jake Wharton
+ */
+public class BasicBus implements Bus {
+  private static final String DEFAULT_IDENTIFIER = "default";
+
+  /** All registered event handlers, indexed by event type. */
+  private final ConcurrentMap<Class<?>, Set<EventHandler>> handlersByType =
+      new ConcurrentHashMap<Class<?>, Set<EventHandler>>();
+
+  /** All registered event producers, index by event type. */
+  private final ConcurrentMap<Class<?>, EventProducer> producersByType =
+      new ConcurrentHashMap<Class<?>, EventProducer>();
+
+  /** A cache of flattened class hierarchies. */
+  private final Map<Class<?>, Set<Class<?>>> flattenedHierarchyCache =
+      new LinkedHashMap<Class<?>, Set<Class<?>>>();
+
+  /** Identifier used to differentiate the event bus instance. */
+  private final String identifier;
+
+  /** Thread enforcer for register, unregister, and posting events. */
+  private final ThreadEnforcer enforcer;
+
+  /** Used to find handler methods in register and unregister. */
+  private final HandlerFinder handlerFinder;
+
+  /** Queues of events for the current thread to dispatch. */
+  private final ThreadLocal<ConcurrentLinkedQueue<EventWithHandler>> eventsToDispatch =
+      new ThreadLocal<ConcurrentLinkedQueue<EventWithHandler>>() {
+        @Override protected ConcurrentLinkedQueue<EventWithHandler> initialValue() {
+          return new ConcurrentLinkedQueue<EventWithHandler>();
+        }
+      };
+
+  /** {@code true} if the current thread is currently dispatching an event. */
+  private final ThreadLocal<Boolean> isDispatching = new ThreadLocal<Boolean>() {
+    @Override protected Boolean initialValue() {
+      return false;
+    }
+  };
+
+  /** Creates a new BasicBus named "default" that enforces actions on the main thread. */
+  public BasicBus() {
+    this(DEFAULT_IDENTIFIER);
+  }
+
+  /**
+   * Creates a new BasicBus with the given {@code identifier} that enforces actions on the main
+   * thread.
+   *
+   * @param identifier a brief name for this bus, for debugging purposes.  Should be a valid Java
+   * identifier.
+   */
+  public BasicBus(String identifier) {
+    this(ThreadEnforcer.MAIN, identifier);
+  }
+
+  /**
+   * Creates a new BasicBus named "default" with the given {@code enforcer} for actions.
+   *
+   * @param enforcer Thread enforcer for register, unregister, and post actions.
+   */
+  public BasicBus(ThreadEnforcer enforcer) {
+    this(enforcer, DEFAULT_IDENTIFIER);
+  }
+
+  /**
+   * Creates a new BasicBus with the given {@code enforcer} for actions and the given {@code
+   * identifier}.
+   *
+   * @param enforcer Thread enforcer for register, unregister, and post actions.
+   * @param identifier A brief name for this bus, for debugging purposes.  Should be a valid Java
+   * identifier.
+   */
+  public BasicBus(ThreadEnforcer enforcer, String identifier) {
+    this(enforcer, identifier, HandlerFinder.ANNOTATED);
+  }
+
+  /**
+   * Test constructor which allows replacing the default {@code HandlerFinder}.
+   *
+   * @param enforcer Thread enforcer for register, unregister, and post actions.
+   * @param identifier A brief name for this bus, for debugging purposes.  Should be a valid Java
+   * identifier.
+   * @param handlerFinder Used to discover event handlers and producers when
+   * registering/unregistering an object.
+   */
+  BasicBus(ThreadEnforcer enforcer, String identifier, HandlerFinder handlerFinder) {
+    if (enforcer == null) {
+      throw new NullPointerException("Enforcer may not be null.");
+    }
+    if (identifier == null) {
+      throw new NullPointerException("Identifier may not be null.");
+    }
+    if (handlerFinder == null) {
+      throw new NullPointerException("Handler finder may not be null.");
+    }
+    this.enforcer = enforcer;
+    this.identifier = identifier;
+    this.handlerFinder = handlerFinder;
+  }
+
+  @Override public String toString() {
+    return "[BasicBus \"" + identifier + "\"]";
+  }
+
+  @Override public void register(Object object) {
+    enforcer.enforce(this);
+
+    Map<Class<?>, EventProducer> foundProducers = handlerFinder.findAllProducers(object);
+    for (Class<?> type : foundProducers.keySet()) {
+
+      final EventProducer producer = foundProducers.get(type);
+      EventProducer previousProducer = producersByType.putIfAbsent(type, producer);
+      //checking if the previous producer existed
+      if (previousProducer != null) {
+        throw new IllegalArgumentException(String.format(
+            "Producer method for type %s found on type %s but already registered by type %s.", type,
+            producer.target.getClass(), previousProducer.target.getClass()));
+      }
+      Set<EventHandler> handlers = handlersByType.get(type);
+      if (handlers != null && !handlers.isEmpty()) {
+        for (EventHandler handler : handlers) {
+          dispatchProducerResultToHandler(handler, producer);
+        }
+      }
+    }
+
+    Map<Class<?>, Set<EventHandler>> foundHandlersMap = handlerFinder.findAllSubscribers(object);
+    for (Class<?> type : foundHandlersMap.keySet()) {
+      Set<EventHandler> handlers = handlersByType.get(type);
+      if (handlers == null) {
+        //concurrent put if absent
+        Set<EventHandler> handlersCreation = new CopyOnWriteArraySet<EventHandler>();
+        handlers = handlersByType.putIfAbsent(type, handlersCreation);
+        if (handlers == null) {
+          handlers = handlersCreation;
+        }
+      }
+      final Set<EventHandler> foundHandlers = foundHandlersMap.get(type);
+      handlers.addAll(foundHandlers);
+    }
+
+    for (Map.Entry<Class<?>, Set<EventHandler>> entry : foundHandlersMap.entrySet()) {
+      Class<?> type = entry.getKey();
+      EventProducer producer = producersByType.get(type);
+      if (producer != null && producer.isValid()) {
+        Set<EventHandler> foundHandlers = entry.getValue();
+        for (EventHandler foundHandler : foundHandlers) {
+          //noinspection ConstantConditions
+          if (!producer.isValid()) {
+            break;
+          }
+          if (foundHandler.isValid()) {
+            dispatchProducerResultToHandler(foundHandler, producer);
+          }
+        }
+      }
+    }
+  }
+
+  private void dispatchProducerResultToHandler(EventHandler handler, EventProducer producer) {
+    Object event = null;
+    try {
+      event = producer.produceEvent();
+    } catch (InvocationTargetException e) {
+      throwRuntimeException("Producer " + producer + " threw an exception.", e);
+    }
+    if (event == null) {
+      return;
+    }
+    dispatch(event, handler);
+  }
+
+  @Override public void unregister(Object object) {
+    enforcer.enforce(this);
+
+    Map<Class<?>, EventProducer> producersInListener = handlerFinder.findAllProducers(object);
+    for (Map.Entry<Class<?>, EventProducer> entry : producersInListener.entrySet()) {
+      final Class<?> key = entry.getKey();
+      EventProducer producer = getProducerForEventType(key);
+      EventProducer value = entry.getValue();
+
+      if (value == null || !value.equals(producer)) {
+        throw new IllegalArgumentException("Missing event producer for an annotated method. Is "
+            + object.getClass()
+            + " registered?");
+      }
+      producersByType.remove(key).invalidate();
+    }
+
+    Map<Class<?>, Set<EventHandler>> handlersInListener = handlerFinder.findAllSubscribers(object);
+    for (Map.Entry<Class<?>, Set<EventHandler>> entry : handlersInListener.entrySet()) {
+      Set<EventHandler> currentHandlers = getHandlersForEventType(entry.getKey());
+      Collection<EventHandler> eventMethodsInListener = entry.getValue();
+
+      if (currentHandlers == null || !currentHandlers.containsAll(eventMethodsInListener)) {
+        throw new IllegalArgumentException("Missing event handler for an annotated method. Is "
+            + object.getClass()
+            + " registered?");
+      }
+
+      for (EventHandler handler : currentHandlers) {
+        if (eventMethodsInListener.contains(handler)) {
+          handler.invalidate();
+        }
+      }
+      currentHandlers.removeAll(eventMethodsInListener);
+    }
+  }
+
+  @Override public void post(Object event) {
+    enforcer.enforce(this);
+
+    Set<Class<?>> dispatchTypes = flattenHierarchy(event.getClass());
+
+    boolean dispatched = false;
+    for (Class<?> eventType : dispatchTypes) {
+      Set<EventHandler> wrappers = getHandlersForEventType(eventType);
+
+      if (wrappers != null && !wrappers.isEmpty()) {
+        dispatched = true;
+        for (EventHandler wrapper : wrappers) {
+          enqueueEvent(event, wrapper);
+        }
+      }
+    }
+
+    if (!dispatched && !(event instanceof DeadEvent)) {
+      post(new DeadEvent(this, event));
+    }
+
+    dispatchQueuedEvents();
+  }
+
+  /**
+   * Queue the {@code event} for dispatch during {@link #dispatchQueuedEvents()}. Events are queued
+   * in-order of occurrence so they can be dispatched in the same order.
+   */
+  protected void enqueueEvent(Object event, EventHandler handler) {
+    eventsToDispatch.get().offer(new EventWithHandler(event, handler));
+  }
+
+  /**
+   * Drain the queue of events to be dispatched. As the queue is being drained, new events may be
+   * posted to the end of the queue.
+   */
+  protected void dispatchQueuedEvents() {
+    // Don't dispatch if we're already dispatching, that would allow reentrancy and out-of-order
+    // events. Instead, leave the events to be dispatched after in-progress dispatch is complete.
+    if (isDispatching.get()) {
+      return;
+    }
+
+    isDispatching.set(true);
+    try {
+      while (true) {
+        EventWithHandler eventWithHandler = eventsToDispatch.get().poll();
+        if (eventWithHandler == null) {
+          break;
+        }
+
+        if (eventWithHandler.handler.isValid()) {
+          dispatch(eventWithHandler.event, eventWithHandler.handler);
+        }
+      }
+    } finally {
+      isDispatching.set(false);
+    }
+  }
+
+  /**
+   * Dispatches {@code event} to the handler in {@code wrapper}.  This method is an appropriate
+   * override point for subclasses that wish to make event delivery asynchronous.
+   *
+   * @param event event to dispatch.
+   * @param wrapper wrapper that will call the handler.
+   */
+  protected void dispatch(Object event, EventHandler wrapper) {
+    try {
+      wrapper.handleEvent(event);
+    } catch (InvocationTargetException e) {
+      throwRuntimeException(
+          String.format("Could not dispatch event %s to handler %s.", event.getClass(), wrapper),
+          e);
+    }
+  }
+
+  /**
+   * Retrieves a mutable set of the currently registered producers for {@code type}.  If no
+   * producers are currently registered for {@code type}, this method will return {@code null}.
+   *
+   * @param type type of producers to retrieve.
+   * @return currently registered producer, or {@code null}.
+   */
+  EventProducer getProducerForEventType(Class<?> type) {
+    return producersByType.get(type);
+  }
+
+  /**
+   * Retrieves a mutable set of the currently registered handlers for {@code type}.  If no handlers
+   * are currently registered for {@code type}, this method may either return {@code null} or an
+   * empty set.
+   *
+   * @param type type of handlers to retrieve.
+   * @return currently registered handlers, or {@code null}.
+   */
+  Set<EventHandler> getHandlersForEventType(Class<?> type) {
+    return handlersByType.get(type);
+  }
+
+  /**
+   * Flattens a class's type hierarchy into a set of Class objects.  The set will include all
+   * superclasses (transitively), and all interfaces implemented by these superclasses.
+   *
+   * @param concreteClass class whose type hierarchy will be retrieved.
+   * @return {@code clazz}'s complete type hierarchy, flattened and uniqued.
+   */
+  Set<Class<?>> flattenHierarchy(Class<?> concreteClass) {
+    Set<Class<?>> classes = flattenedHierarchyCache.get(concreteClass);
+    if (classes == null) {
+      classes = getClassesFor(concreteClass);
+      flattenedHierarchyCache.put(concreteClass, classes);
+    }
+
+    return classes;
+  }
+
+  private Set<Class<?>> getClassesFor(Class<?> concreteClass) {
+    List<Class<?>> parents = new LinkedList<Class<?>>();
+    Set<Class<?>> classes = new HashSet<Class<?>>();
+
+    parents.add(concreteClass);
+
+    while (!parents.isEmpty()) {
+      Class<?> clazz = parents.remove(0);
+      classes.add(clazz);
+
+      Class<?> parent = clazz.getSuperclass();
+      if (parent != null) {
+        parents.add(parent);
+      }
+    }
+    return classes;
+  }
+
+  /**
+   * Throw a {@link RuntimeException} with given message and cause lifted from an {@link
+   * InvocationTargetException}. If the specified {@link InvocationTargetException} does not have a
+   * cause, neither will the {@link RuntimeException}.
+   */
+  private static void throwRuntimeException(String msg, InvocationTargetException e) {
+    Throwable cause = e.getCause();
+    if (cause != null) {
+      throw new RuntimeException(msg, cause);
+    } else {
+      throw new RuntimeException(msg);
+    }
+  }
+
+  /** Simple struct representing an event and its handler. */
+  static class EventWithHandler {
+    final Object event;
+    final EventHandler handler;
+
+    public EventWithHandler(Object event, EventHandler handler) {
+      this.event = event;
+      this.handler = handler;
+    }
+  }
+}
