+++ library/src/main/java/com/squareup/otto/Produce.java
@@ -22,14 +22,15 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Marks a method as an instance producer, as used by {@link AnnotatedHandlerFinder} and {@link Bus}.
+ * Marks a method as an instance producer, as used by {@link AnnotatedHandlerFinder} and {@link
+ * BasicBus}.
  * <p>
- * Otto infers the instance type from the annotated method's return type. Producer methods may return null when there is
- * no appropriate value to share. The calling {@link Bus} ignores such returns and posts nothing.
+ * Otto infers the instance type from the annotated method's return type. Producer methods may
+ * return null when there is no appropriate value to share. The calling {@link BasicBus} ignores
+ * such returns and posts nothing.
  *
  * @author Jake Wharton
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME) @Target(ElementType.METHOD)
 public @interface Produce {
 }
