+++ library/src/main/java/com/squareup/otto/Subscribe.java
@@ -22,16 +22,17 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Marks a method as an event handler, as used by {@link AnnotatedHandlerFinder} and {@link Bus}.
- *
- * <p>The method's first (and only) parameter defines the event type.
- * <p>If this annotation is applied to methods with zero parameters or more than one parameter, the object containing
- * the method will not be able to register for event delivery from the {@link Bus}. Otto fails fast by throwing
- * runtime exceptions in these cases.
+ * Marks a method as an event handler, as used by {@link AnnotatedHandlerFinder} and {@link
+ * BasicBus}.
+ * <p>
+ * The method's first (and only) parameter defines the event type.
+ * <p>
+ * If this annotation is applied to methods with zero parameters or more than one parameter, the
+ * object containing the method will not be able to register for event delivery from the {@link
+ * BasicBus}. Otto fails fast by throwing runtime exceptions in these cases.
  *
  * @author Cliff Biffle
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME) @Target(ElementType.METHOD)
 public @interface Subscribe {
 }
