+++ sample/pom.xml
@@ -20,11 +20,10 @@
   <parent>
     <groupId>com.squareup</groupId>
     <artifactId>otto-parent</artifactId>
-    <version>1.3.4-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
-  <groupId>com.squareup</groupId>
   <artifactId>otto-sample</artifactId>
   <packaging>apk</packaging>
   <name>Otto Sample</name>
@@ -47,23 +46,11 @@
   </dependencies>
 
   <build>
-    <finalName>${binary.prefix}-${project.artifactId}-${project.version}</finalName>
-
     <plugins>
       <plugin>
         <groupId>com.jayway.maven.plugins.android.generation2</groupId>
         <artifactId>android-maven-plugin</artifactId>
         <extensions>true</extensions>
-      </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-      </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-checkstyle-plugin</artifactId>
       </plugin>
     </plugins>
   </build>
