+++ sample/src/main/java/com/squareup/otto/sample/BusProvider.java
@@ -16,14 +16,15 @@
 
 package com.squareup.otto.sample;
 
+import com.squareup.otto.BasicBus;
 import com.squareup.otto.Bus;
 
 /**
- * Maintains a singleton instance for obtaining the bus. Ideally this would be replaced with a more efficient means
- * such as through injection directly into interested classes.
+ * Maintains a singleton instance for obtaining the bus. Ideally this would be replaced with a more
+ * efficient means such as through injection directly into interested classes.
  */
 public final class BusProvider {
-  private static final Bus BUS = new Bus();
+  private static final Bus BUS = new BasicBus();
 
   public static Bus getInstance() {
     return BUS;
